@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
 part 'walletaddress.g.dart';
 
@@ -9,24 +8,24 @@ class WalletAddress extends HiveObject {
   @HiveField(1)
   String addressBookName;
   @HiveField(2)
-  String status;
+  String? status;
   @HiveField(3)
   bool used;
   @HiveField(4)
-  bool isOurs = true;
+  bool? isOurs = true; //nullable for backward compatability
   @HiveField(5)
-  String wif = '';
+  String? wif = ''; //nullable for backward compatability
 
   WalletAddress({
-    @required this.address,
-    @required this.addressBookName,
-    @required this.used,
-    @required this.status,
-    @required this.isOurs,
-    @required this.wif,
+    required this.address,
+    required this.addressBookName,
+    required this.used,
+    required this.status,
+    required this.isOurs,
+    required this.wif,
   });
 
-  set newStatus(String newStatus) {
+  set newStatus(String? newStatus) {
     status = newStatus;
   }
 
