@@ -16,12 +16,12 @@ class WalletImportScanScreen extends StatefulWidget {
 class _WalletImportScanScreenState extends State<WalletImportScanScreen> {
   bool _initial = true;
   bool _scanStarted = false;
-  ElectrumConnection _connectionProvider;
-  ActiveWallets _activeWallets;
-  String _coinName = '';
-  ElectrumConnectionState _connectionState;
+  ElectrumConnection? _connectionProvider;
+  late ActiveWallets _activeWallets;
+  late String _coinName = '';
+  late ElectrumConnectionState _connectionState;
   int _latestUpdate = 0;
-  Timer _timer;
+  late Timer _timer;
 
   @override
   void didChangeDependencies() async {
@@ -29,23 +29,23 @@ class _WalletImportScanScreenState extends State<WalletImportScanScreen> {
       setState(() {
         _initial = false;
       });
-      _coinName = ModalRoute.of(context).settings.arguments as String;
+      _coinName = ModalRoute.of(context)!.settings.arguments as String;
       _connectionProvider = Provider.of<ElectrumConnection>(context);
       _activeWallets = Provider.of<ActiveWallets>(context);
       await _activeWallets.prepareForRescan(_coinName);
-      await _connectionProvider.init(_coinName, scanMode: true);
+      await _connectionProvider!.init(_coinName, scanMode: true);
 
       _timer = Timer.periodic(Duration(seconds: 5), (timer) async {
         var dueTime = _latestUpdate + 5;
         if (_connectionState == ElectrumConnectionState.waiting) {
-          await _connectionProvider.init(_coinName, scanMode: true);
+          await _connectionProvider!.init(_coinName, scanMode: true);
         } else if (dueTime <= DateTime.now().millisecondsSinceEpoch ~/ 1000) {
           _timer.cancel();
           await Navigator.of(context).pushReplacementNamed(Routes.WalletList);
         }
       });
     } else if (_connectionProvider != null) {
-      _connectionState = _connectionProvider.connectionState;
+      _connectionState = _connectionProvider!.connectionState;
 
       if (_connectionState == ElectrumConnectionState.connected) {
         _latestUpdate = DateTime.now().millisecondsSinceEpoch ~/ 1000;
@@ -56,7 +56,7 @@ class _WalletImportScanScreenState extends State<WalletImportScanScreen> {
               _coinName, 0, 0, 0, true);
 
           //subscribe to hd master
-          _connectionProvider.subscribeToScriptHashes(await _activeWallets
+          _connectionProvider!.subscribeToScriptHashes(await _activeWallets
               .getWalletScriptHashes(_coinName, _masterAddr));
           setState(() {
             _scanStarted = true;
@@ -76,7 +76,7 @@ class _WalletImportScanScreenState extends State<WalletImportScanScreen> {
 
   @override
   void deactivate() async {
-    await _connectionProvider.closeConnection();
+    await _connectionProvider!.closeConnection();
     _timer.cancel();
     super.deactivate();
   }
