@@ -5,7 +5,7 @@ part 'app_options.g.dart';
 class AppOptionsStore extends HiveObject {
   @HiveField(0)
   // ignore: prefer_final_fields
-  Map<String, bool> _authenticationOptions = {
+  Map<String, bool>? _authenticationOptions = {
     'walletList': false,
     'walletHome': false,
     'sendTransaction': true,
@@ -13,15 +13,15 @@ class AppOptionsStore extends HiveObject {
   };
 
   @HiveField(1)
-  bool _allowBiometrics = true;
+  bool? _allowBiometrics = true;
 
   @HiveField(2)
-  String _defaultWallet = '';
+  String? _defaultWallet = '';
 
   AppOptionsStore(this._allowBiometrics);
 
   bool get allowBiometrics {
-    return _allowBiometrics;
+    return _allowBiometrics ?? false;
   }
 
   set allowBiometrics(bool newStatus) {
@@ -29,17 +29,17 @@ class AppOptionsStore extends HiveObject {
     save();
   }
 
-  Map<String, bool> get authenticationOptions {
+  Map<String, bool>? get authenticationOptions {
     return _authenticationOptions;
   }
 
   void changeAuthenticationOptions(String field, bool value) {
-    _authenticationOptions[field] = value;
+    _authenticationOptions![field] = value;
     save();
   }
 
   String get defaultWallet {
-    return _defaultWallet;
+    return _defaultWallet ?? '';
   }
 
   set defaultWallet(String newWallet) {
