@@ -24,8 +24,10 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
 
   Future<void> addWallet(ctx) async {
     try {
-      await Provider.of<ActiveWallets>(context, listen: false).addWallet(_coin,
-          availableCoins[_coin].displayName, availableCoins[_coin].letterCode);
+      await Provider.of<ActiveWallets>(context, listen: false).addWallet(
+          _coin,
+          availableCoins[_coin]!.displayName,
+          availableCoins[_coin]!.letterCode);
       var prefs =
           await Provider.of<UnencryptedOptions>(context, listen: false).prefs;
 
@@ -53,7 +55,7 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
   void didChangeDependencies() async {
     if (_initial) {
       var _appSettings = Provider.of<AppSettings>(context, listen: false);
-      if (_appSettings.authenticationOptions['newWallet']) {
+      if (_appSettings.authenticationOptions!['newWallet']!) {
         await Auth.requireAuth(context, _appSettings.biometricsAllowed);
       }
       setState(() {
@@ -88,6 +90,7 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
             Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 10.0),
                 child: IconButton(
+                  key: Key('confirm_add_wallet'),
                   onPressed: () => addWallet(context),
                   icon: Icon(Icons.add),
                 ))
@@ -119,18 +122,18 @@ class _NewWalletScreenState extends State<NewWalletScreen> {
                               child: Image.asset(
                                   AvailableCoins()
                                       .getSpecificCoin(availableCoins[
-                                              actualAvailableWallets[item]]
+                                              actualAvailableWallets[item]]!
                                           .name)
                                       .iconPath,
                                   width: 20),
                             ),
                             title: Text(
-                                availableCoins[actualAvailableWallets[item]]
+                                availableCoins[actualAvailableWallets[item]]!
                                     .displayName),
                             leading: Radio(
                               value: actualAvailableWallets[item],
                               groupValue: _coin,
-                              onChanged: (value) {
+                              onChanged: (dynamic value) {
                                 setState(
                                   () {
                                     _coin = value;
