@@ -87,7 +87,7 @@ class SettingsAuth extends StatelessWidget {
             title: Text(
               AppLocalizations.instance.translate('app_settings_walletList'),
             ),
-            value: _authenticationOptions['walletList'],
+            value: _authenticationOptions['walletList']!,
             onChanged: (newState) {
               _settings.setAuthenticationOptions('walletList', newState);
               _saveSnack(context);
@@ -96,7 +96,7 @@ class SettingsAuth extends StatelessWidget {
             title: Text(
               AppLocalizations.instance.translate('app_settings_walletHome'),
             ),
-            value: _authenticationOptions['walletHome'],
+            value: _authenticationOptions['walletHome']!,
             onChanged: (newState) {
               _settings.setAuthenticationOptions('walletHome', newState);
               _saveSnack(context);
@@ -106,7 +106,7 @@ class SettingsAuth extends StatelessWidget {
               AppLocalizations.instance
                   .translate('app_settings_sendTransaction'),
             ),
-            value: _authenticationOptions['sendTransaction'],
+            value: _authenticationOptions['sendTransaction']!,
             onChanged: (newState) {
               _settings.setAuthenticationOptions('sendTransaction', newState);
               _saveSnack(context);
@@ -115,7 +115,7 @@ class SettingsAuth extends StatelessWidget {
             title: Text(
               AppLocalizations.instance.translate('app_settings_newWallet'),
             ),
-            value: _authenticationOptions['newWallet'],
+            value: _authenticationOptions['newWallet']!,
             onChanged: (newState) {
               _settings.setAuthenticationOptions('newWallet', newState);
               _saveSnack(context);
